Pass address guid when submitting a bill.
Add method to retrieve electric bills.

Add method to store electric bill

Add query to check if an address exists and if it is associated with the given user id.

+++ ElectricDashboardApi.Data/Entities/ElectricBill.cs
@@ -30,10 +30,10 @@ public class ElectricBill
 
     [Required]
     [Column("consumptionkwh", TypeName = "decimal(10, 2)")]
-    public decimal ConsumptionKwh { get; set; }
+    public int ConsumptionKwh { get; set; }
 
     [Column("sendbackkwh", TypeName = "decimal(10, 2)")]
-    public decimal? SentBackKwh { get; set; }
+    public int? SentBackKwh { get; set; }
 
     [Required]
     [Column("billedamount", TypeName = "decimal(10, 2)")]
